requirements.txt
numpy
scipy
webcolors
pillow
